Unify analyze + diff HTTP routing under /analyses with mode discriminator
The processor unification earlier in this branch made full + diff
share one pipeline, but the HTTP surface was still split across two
inconsistent endpoints — /analyze + /analyze/cancel vs /diff-check
(POST + GET). Collapse to one noun (analyses) with `mode` in the
POST body.

Routes:
  POST   /analyses        body: { mode: 'full'|'diff', skipGit? }
  POST   /analyses/cancel
  GET    /analyses                — history list (unchanged)
  GET    /analyses/diff           — current diff.json (replaces /diff-check)
  GET    /analyses/:id/usage      — unchanged
  DELETE /analyses/:id            — unchanged

Schema change:
AnalyzeRepoSchema shrinks to `{mode, skipGit?}`. Dropped
`enabledCategories` and `enableLlmRules` — both are duplicated by the
per-repo config (`truecourse rules categories/llm`) which is the
canonical source, the web client never sets them, and per-run LLM
opt-out is already covered by the mid-run `analysis:llm-estimate`
socket prompt. `skipGit` stays — no config equivalent.

Implementation:
- routes/analyze.ts deleted; everything folded into routes/analyses.ts.
- POST /analyses dispatches by mode internally — shared prelude
  (register, 202, tracker, createSocketLlmEstimateHandler, pushLogger,
  try/catch with emits, unregister in finally), two small mode-specific
  bodies (runFullAnalyze / runDiffAnalyze).
- index.ts mount cleaned up.
- apps/web/src/lib/api.ts: analyzeRepo / runDiffCheck / getDiffCheck /
  cancelAnalysis repoint at the new paths and bodies.
- tests/shared/schemas.test.ts updated for the new schema shape.

No CLI changes — CLI runs analyze/diff in-process, never hits HTTP.
